Add team endpoints and add team based permissions
Linked: #6

+++ api/lib/time_manager_web/controllers/team_controller.ex
@@ -3,8 +3,68 @@ defmodule TimeManagerWeb.TeamController do
 
   alias TimeManager.Teams
   alias TimeManager.Teams.Team
+  alias TimeManagerWeb.Plugs.CheckPermissions
 
   action_fallback TimeManagerWeb.FallbackController
+
+  plug(CheckPermissions,
+  actions: [
+    clock: %{"team" => ["clock"]},
+    add_employee: %{"team" => ["update"]},
+    remove_employee: %{"team" => ["update"]},
+    add_manager: %{"team" => ["update"]},
+    remove_manager: %{"team" => ["update"]},
+    index: %{"team" => ["read"]},
+    show: %{"team" => ["read"]},
+    create: %{"team" => ["create"]},
+    update: %{"team" => ["update"]},
+    delete: %{"team" => ["delete"]}
+  ]
+)
+
+  def clock(conn, %{"teamID" => id}) do
+    team = TimeManager.Repo.preload(Teams.get_team!(id), :employees)
+    CheckPermissions.assert_team_permissions(conn, id)
+
+    conn
+    |> json(Enum.map(team.employees, fn employee ->
+      with {:ok, %TimeManager.Clocks.Clock{} = clock} <- TimeManager.Clocks.clock_user(employee) do
+        conn
+        |> put_status(:created)
+        |> json(TimeManagerWeb.ClockJSON.show(%{clock: clock}))
+      end
+    end))
+  end
+
+  def add_employee(conn, %{"teamID" => team_id, "employee_id" => employee_id}) do
+    team = Teams.get_team!(team_id)
+    CheckPermissions.assert_team_permissions(conn, team_id)
+
+    Teams.add_employee(team.id, employee_id)
+    render(conn, :show, team: team)
+  end
+
+  def remove_employee(conn, %{"teamID" => team_id, "employee_id" => employee_id}) do
+    team = Teams.get_team!(team_id)
+    CheckPermissions.assert_team_permissions(conn, team_id)
+
+    Teams.remove_employee(team.id, employee_id)
+    render(conn, :show, team: team)
+  end
+
+  def add_manager(conn, %{"teamID" => team_id, "manager_id" => manager_id}) do
+    team = Teams.get_team!(team_id)
+    Teams.add_manager(team.id, manager_id)
+    render(conn, :show, team: team)
+  end
+
+  def remove_manager(conn, %{"teamID" => team_id, "manager_id" => manager_id}) do
+    team = Teams.get_team!(team_id)
+    CheckPermissions.assert_team_permissions(conn, team_id)
+
+    Teams.remove_manager(team.id, manager_id)
+    render(conn, :show, team: team)
+  end
 
   def index(conn, _params) do
     teams = Teams.list_teams()
@@ -12,7 +72,10 @@ defmodule TimeManagerWeb.TeamController do
   end
 
   def create(conn, %{"team" => team_params}) do
+    CheckPermissions.assert_manager(conn)
+
     with {:ok, %Team{} = team} <- Teams.create_team(team_params) do
+      Teams.add_manager(team.id, TimeManagerWeb.Plugs.CheckPermissions.get_user_id(conn))
       conn
       |> put_status(:created)
       |> put_resp_header("location", ~p"/api/teams/#{team.id}")
@@ -22,11 +85,14 @@ defmodule TimeManagerWeb.TeamController do
 
   def show(conn, %{"id" => id}) do
     team = Teams.get_team!(id)
+    CheckPermissions.assert_team_read_permissions(conn, id)
+
     render(conn, :show, team: team)
   end
 
   def update(conn, %{"id" => id, "team" => team_params}) do
     team = Teams.get_team!(id)
+    CheckPermissions.assert_team_permissions(conn, id)
 
     with {:ok, %Team{} = team} <- Teams.update_team(team, team_params) do
       render(conn, :show, team: team)
@@ -35,6 +101,7 @@ defmodule TimeManagerWeb.TeamController do
 
   def delete(conn, %{"id" => id}) do
     team = Teams.get_team!(id)
+    CheckPermissions.assert_team_permissions(conn, id)
 
     with {:ok, %Team{}} <- Teams.delete_team(team) do
       send_resp(conn, :no_content, "")
